Agrego cosas

+++ src/main/java/ar/com/sge/usuarios/Cliente.java
@@ -47,8 +47,8 @@ public class Cliente extends Usuario {
 		this.transformador = transformador;
 	}
 
-	public Cliente(String _nombre, String _apellido, String tipoDoc, int numeroDoc, int telefono, Categoria categoria,int puntos,float latitud,float longitud ) {
-		super(_nombre, _apellido,latitud,longitud);
+	public Cliente(String _nombre, String _apellido, String tipoDoc, int numeroDoc, int telefono, Categoria categoria,int puntos,double latitud,double longitud ) {
+		super(_nombre,_apellido,latitud,longitud);
 		this.tipoDoc = tipoDoc;
 		this.numeroDoc = numeroDoc;
 		this.telefono = telefono;
